+++ codex-rs/mcp-server/src/lib.rs
@@ -6,10 +6,10 @@ use std::io::Result as IoResult;
 use std::path::PathBuf;
 
 use codex_common::CliConfigOverrides;
+use codex_core::config::AdminAuditContext;
 use codex_core::config::Config;
 use codex_core::config::ConfigOverrides;
-use codex_core::config::audit_admin_run_with_prompt;
-use codex_core::config::prompt_for_admin_danger_reason;
+use codex_core::config::audit_admin_run_without_prompt;
 
 use mcp_types::JSONRPCMessage;
 use tokio::io::AsyncBufReadExt;
@@ -98,7 +98,20 @@ pub async fn run_main(
         .map_err(|e| {
             std::io::Error::new(ErrorKind::InvalidData, format!("error loading config: {e}"))
         })?;
-    audit_admin_run_with_prompt(&config, prompt_for_admin_danger_reason(&config), false).await?;
+
+    audit_admin_run_without_prompt(
+        &config.admin_controls,
+        AdminAuditContext {
+            sandbox_policy: &config.sandbox_policy,
+            approval_policy: &config.approval_policy,
+            cwd: &config.cwd,
+            command: None,
+            dangerously_bypass_requested: false,
+            dangerous_mode_justification: None,
+            record_command_event: false,
+        },
+    )
+    .await?;
 
     // Task: process incoming messages.
     let processor_handle = tokio::spawn({
